requirement.txt
langchain
pypdf
qdrant_client
openai
streamlit
python-dotenv